refactor : ScheduleCloseEventHandler에 타이틀 분석 로직 추가
